🥅 Catch syntax errors when loading settings
Prior to this commit, any errors when trying to load a settings file were simply
ignored and the file was overwritten with an empty one. Now the error is
displayed to the user and the program exits.
The only time a new file is created is when the settings file does not already
exist.