♻️ Refactor rendering resource management
With the Lua thread separate from the render thread now it is possible
that rendering resources, notably textures, can be destroyed while a
frame is being rendered.

In the process of resolving the above, I also found that the frame swap
logic was wrong, it was waiting for the frame to finish rendering before
continuing to the next; this was essentially single buffered instead of
double.

Fixing the above made the first issue even worse: even after a frame has
ended, the GPU may still be executing the commands.

To resolve this, I've introduced resource referencing for resources,
including textures and VBOs, and pipeline states. Each backbuffer has a
set of resource references that get released only when the frame is
verified to be finished on the GPU.